requirements.txt
tensorflow
tensorflow
tensorflow_hub
python-multipart
pillow
matplotlib
numpy
streamlit